p4i integration Integrate 216259 from 4.1-optimized to main: Integrate 216258 from 4.1 to 4.1-optimized: Fixes:    QPicture not painting sometimes Task:     #108599 RevBy:    Eskil Details:  The integer division here can yield zero under certain circumstances                    (think device->logicalDpiX() == 96, qt_defaultDpi() == 95), causing a                    badly scaled world matrix. Divide with floating point precision instead.
[git-p4: depot-paths = "//depot/qt/main/": change = 216260]

+++ src/gui/image/qpicture.cpp
@@ -453,8 +453,8 @@ bool QPicture::exec(QPainter *painter, QDataStream &s, int nrecords)
 
     QMatrix worldMatrix = painter->matrix();
     QMatrix oldWorldMatrix = worldMatrix;
-    worldMatrix.scale(painter->device()->logicalDpiX() / qt_defaultDpi(),
-                      painter->device()->logicalDpiY() / qt_defaultDpi());
+    worldMatrix.scale(qreal(painter->device()->logicalDpiX()) / qreal(qt_defaultDpi()),
+                      qreal(painter->device()->logicalDpiY()) / qreal(qt_defaultDpi()));
     painter->setMatrix(worldMatrix);
 
     while (nrecords-- && !s.atEnd()) {
